// generated/src/app/api-models/cap-stoploss-select.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapStoplossSelect {
  capModelId: string;
  capStoplossId: string;
  seqStoplossSel: number;
  columnName: string;
  columnOccurrence: number;
  fromValue: string;
  thruValue: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  operator: string;
  columnType: string;
}